refactor: move bitswap stats directly under the Bitswap object
