+++ bitswap/src/ledger.rs
@@ -8,62 +8,9 @@ use prost::Message as ProstMessage;
 use std::{
     collections::{HashMap, HashSet},
     mem,
-    sync::{
-        atomic::{AtomicU64, Ordering},
-        Arc,
-    },
 };
 
 pub type Priority = i32;
-
-#[derive(Debug, Default)]
-pub struct Stats {
-    pub sent_blocks: AtomicU64,
-    pub sent_data: AtomicU64,
-    pub received_blocks: AtomicU64,
-    pub received_data: AtomicU64,
-    pub duplicate_blocks: AtomicU64,
-    pub duplicate_data: AtomicU64,
-}
-
-impl Stats {
-    pub fn update_outgoing(&self, num_blocks: u64) {
-        self.sent_blocks.fetch_add(num_blocks, Ordering::Relaxed);
-    }
-
-    pub fn update_incoming_unique(&self, bytes: u64) {
-        self.received_blocks.fetch_add(1, Ordering::Relaxed);
-        self.received_data.fetch_add(bytes, Ordering::Relaxed);
-    }
-
-    pub fn update_incoming_duplicate(&self, bytes: u64) {
-        self.duplicate_blocks.fetch_add(1, Ordering::Relaxed);
-        self.duplicate_data.fetch_add(bytes, Ordering::Relaxed);
-    }
-
-    pub fn add_assign(&self, other: &Stats) {
-        self.sent_blocks
-            .fetch_add(other.sent_blocks.load(Ordering::Relaxed), Ordering::Relaxed);
-        self.sent_data
-            .fetch_add(other.sent_data.load(Ordering::Relaxed), Ordering::Relaxed);
-        self.received_blocks.fetch_add(
-            other.received_blocks.load(Ordering::Relaxed),
-            Ordering::Relaxed,
-        );
-        self.received_data.fetch_add(
-            other.received_data.load(Ordering::Relaxed),
-            Ordering::Relaxed,
-        );
-        self.duplicate_blocks.fetch_add(
-            other.duplicate_blocks.load(Ordering::Relaxed),
-            Ordering::Relaxed,
-        );
-        self.duplicate_data.fetch_add(
-            other.duplicate_data.load(Ordering::Relaxed),
-            Ordering::Relaxed,
-        );
-    }
-}
 
 /// The Ledger contains the history of transactions with a peer.
 #[derive(Debug, Default)]
@@ -74,8 +21,6 @@ pub struct Ledger {
     pub(crate) received_want_list: HashMap<Cid, Priority>,
     /// Queued message.
     message: Message,
-    /// Statistics related to a given peer.
-    pub stats: Arc<Stats>,
 }
 
 impl Ledger {
@@ -115,8 +60,6 @@ impl Ledger {
         for (cid, priority) in self.message.want() {
             self.sent_want_list.insert(cid.clone(), *priority);
         }
-
-        self.stats.update_outgoing(self.message.blocks.len() as u64);
 
         Some(mem::take(&mut self.message))
     }
