+++ bitswap/src/lib.rs
@@ -9,10 +9,10 @@ mod ledger;
 mod prefix;
 mod protocol;
 
-pub use self::behaviour::{Bitswap, BitswapEvent};
+pub use self::behaviour::{Bitswap, BitswapEvent, Stats};
 pub use self::block::Block;
 pub use self::error::BitswapError;
-pub use self::ledger::{Priority, Stats};
+pub use self::ledger::Priority;
 
 mod bitswap_pb {
     include!(concat!(env!("OUT_DIR"), "/bitswap_pb.rs"));
